requirements.txt
segno~=1.6.1
streamlit~=1.36.0
requests~=2.32.3
pillow~=10.3.0